fix: componentização da seção de informações de compra do produto
